Allows unauth access; Add data sources and dashboards to Grafana
This is complicated by the fact that Grafana sets data sources and
dashboards in an internal sqlite3 database.  The easiest work around
was to simply extract this database, and add it to the Docker image.

In the future, we should build the image based on a set of exported
dashboard JSON files.  There doesn't seem to be much of a work-around
for the data sources, but they change infrequently, so it is somewhat
less of an issue.